Check src->obs calculation for a delay operator *only*
This implementation seems to work; there was some weird stuff going on with
the start index prior to this, and I think I was interpolating *forward* in
time instead of backwards. You can do this if the particles are well-separated
(and I think my test cases had them well-separated) but it's best to avoid
this all-together.