"Fix" minus sign error in 1st derivative coefficients
I don't know why this minus sign needs to be here and the coefficients
match what I generate in Mathematica without it, but evaluating a function's
derivative without it pretty clearly produces the wrong result. So here's
a minus sign and a comment until I figure out a better explanation.