Construct SoA "InteractionTable"
I anticipate it'll be far faster to build a contiguous interaction table
than it will be to construct a table of Interaction objects (each with their
own non-memory-local tables) simply because of the reduced overhead in
allocating one thing vs a billion. This is my first go at that monolithic
SoA design, but the numbers agree with those computed in a straight Interaction.
Whether or not they're physically right has yet to be determined, though.