Use assistant builder class to initialize PredictorCorrectors
Before, the P/C declared a BUNCH of stuff that it needed to initialize
itself but didn't need afterwards. This garbage stuck around, however,
until the P/C itself went out of scope (most likely at the end of main()).
Now, the P/C constructor offloads a bunch of this work into a Builder class
that creates some data, shares it around to create the P/C coefficients,
and then returns them to the *actual* predictor corrector. Much cleaner,
particularly because I want the actual predictor/corrector to manage the
time history of the quantum dots.