Shorten up names of coefficient tables
Ordinarily I'm not a fan of short, two-letter variable names. I've done this
in the process of refactoring the integration routines to consolidate their
behavior (an Integrator holds a Weights object and a .step() method) so
there should be a *ton* of additional context around .ps() and .cs() (the
predictor and corrector tables) making the new variable names very clear.