Imbue Integrator with shared_ptr to HistoryArray
With the change to a HistoryInteraction sort of scheme, many different parts
of the code require knowledge of the system's history. Thus, everyone gets
a shared pointer to what stands in for a history object and the integration
required retooling accordingly. Here it's untested, but hopefully soon
everything will be segregated enough to facilitate easy tests of *all*
modules.