Retool integrator to get bounds History object
Previously, I had kind of thought of the Integrator as controlling the
History uniquely. Other things (Interactions) need access to the history,
too, so it needs to be a shared object (i.e., constructed externally and
passed to the Integrator in its constructor). Because the Integrator doesn't
built the table itself, it doesn't need to explicitly have information about
the "number of timesteps" or "number of solutions"; it can figure them all
out from the History object that gets passed in.