Include point generation program
QuEST pretty much requires a minimum distance between points to work
well, but checking the minimum distance between 10000+ particles is slow
work in Mathematica or Python, so I opted for a blazing C++ version.
Little harder to generalize to arbitrary distributions, but the speedup
in checking distances is worth it.