Overhaul cmake files
The article at https://crascit.com/2016/01/31/enhanced-source-file-handling-with-target_sources/
makes a good argument for using target_sources() in cmake. This has the
advantage of making lots of links in the compile "chain" explicit, but
it's a little trickier than just globbing everything (as was done previously).
Hopefully this will be more amenable to unit testing in the future, too.