Rejigger cmake to build libquest and quest(.exe)
All of the compiled object files will need to link in to power the unit test
sources as well, so it makes sense to pull them into a library (libquest)
and compile them just once. This should do just that, and then use libquest
to link with the quest executable.