Retool bounds calculation
The previous code tried to accomodate planar or other "reduced" geometries
by adding 1 to each dimension's shape. This wasn't quite right because of
how sources-expanding-into-grid-points works for AIM (because floor(pt) and
ceil(pt) both need to exist "in the grid"), so it required a bit of fixing.
Should work much better now with a more appropriate ubound and an optional
padding argument for higher-order expansions.