Pull out (real) gmatrix construction into separate function
Because of how FFTW cannibalizes the input array to do its testing, the
order in which arrays need to get allocated and filled is complicated. Using
a void function with a reference argument to fill out the gmatrix table
(before fft'ing) makes the most sense, I think. It makes it easy to generate
a table and check its contents separately from the fft function and separates
the concerns a little bit better.