Remove the box-point mapping member of Grids
Wasn't really appropriate for the grid to manage this information, so it's
become a function that returns the appropriate object (instead of a function
that fills up an internal data member). This is to help separate grid
construction (which requires a mutable dot vector for sorting) and AIM
construction (which does not).