Split HistoryInteraction into ABC + DirectInteraction
The new AIM stuff in the pipeline requires the exact same functionality
as the current "HistoryInteraction" (for it is itself a history-dependent
interaction). Consequently, it makes sense that *both* the AimInteraction
and (now) DirectInteraction inherit from a HistoryInteraction containing
the common data & behavior. This commit pulls out those commonalities into
an abstract-base HistoryInteraction and moves whatever else was in there
into the new DirectInteraction.