Implement Fourier Index calculation
To reduce the storage cost of the propagation matrices, I only store the
(FFTs) of the first row (nx*ny Fourier transforms, each of length nz).
The rub then becomes how to index into this structure for performing the
actual matrix-vector product. These routines should assist with that;
they provide the mapping from sub-sub-blocks of G to the actual stored
data.