Add a concrete constraint on the fast multiplication
It's _really_ difficult to numerically compare the fast multiply with any
sort of "analytic" calculation as it would require an enormous amount of
additional code work to generate the full g matrices and then take an
inner product. This is done easily enough in Python, and so the recent
"AIM matrix elements.ipynb" should offer a good-enough manual comparison
for the time being.