3d-ify and clean up the spatial transforms
It's much better for everyone and everything to use a 3d transform that
ultimately exploits translational invariance in three dimensions (instead
of just one as the old code did). This commit swaps the transforms to 3d
ones that operate on individual "time blocks" (except when building the
whole gmatrix -- that FFTW call does the whole vector in one go for speed)
and tests them against values computed in Mathematica.