Add on-grid expansion checks
This commit also cleans up the names in expansion_test a bit.

Fix debug-mode FIXTURE error (???)

Somehow when this was compiled in DEBUG mode, the top-level fixture caused
a large number of issues (the box_expansion_order wasn't propagating through
all of the tests correctly). Suuuuuuuuper frustrating, but moving the fixture
from SUITE level to TEST level seems to have fixed it.