Consolidate "new" expansion interface
I needed to tweak the move-able, OO expansion calculation to accommodate
(possible) time derivatives of a sent signal. Ultimately this meant that
the expansion_function needed knowledge of the timestep at which to calculate
its expansion (so it can walk backwards for interpolated time derivatives
if necessary), meaning that the function signatures changed up a bit. This
commit consolidates that interface to re-enable the GAUSSIAN_POINT_PROPAGATION
(and similar) tests.