Remove on-knot value tests
As coded, the lagrange interpolation module cannot accommodate this without
some serious conditional logic to eliminate zeroes that occur in both
the numerator and denominator of the polynomial fractions. Consequently,
any of the time-interpolations that require values at t = 0 (e.g. to conjure
derivatives in the EFIE) will have to implement a workaround until a better
solution presents itself.