Consolidate point-to-point test structure
A lot of the numerical correctness tests for two points interacting through
AIM had nearly identical structures, save for initializing the geometry
of the source and observation points. I _could_ have just copy-pasted
everything to duplicate the test functionality, but that's bad (for obvious
reasons) and this keeps the actual tests pretty concise. Could stand to have
more documentation, though...