Add "experiments" directory + propagation experiments
Grabbing the data necessary for a (numerics) paper is cumbersome with software
that isn't purpose-built to dump out the numerics data. So, this commit
introduces some software that's purpose-built to do just that. It adds
a "propagation_test" executable to generate input and output data (analytic
and simulated) for different propagation kernels so that their errors can
be recorded and indcluded in publication (THESIS!) figures.