Provide RetardationBase
Some of the time modulos are common to a lot of the Expansion things, so
it makes a little sense to pull that functionality into a base class. To
remain generic, the calling code expects a std::function so the derived
classes DO NOT need to overload an operator(); they can just provide their
own. This is probably a little bit faster than doing full-tilt polymorphism
(see: https://stackoverflow.com/questions/16435549/when-to-use-stdfunction-instead-of-inheritance)