Swap to out-of-place FFTs
Doing the nearfield calculations requires un-transformed data on the grid.
While this is available at the current timestep (if the calculation is done
between the to-grid and grid-grid steps if the FFT is pushed into the grid-
grid step, obviously), an in-place transform wrecks the tables for *previous*
timesteps. Making the grid-grid calculation occur from/to separate tables
(named with an _fft suffix) should fix this problem (and will be faster, to
boot!)