Perform FFTs in-place
Per commit c1c8f816, the contiguous source-history vector gets updated in
evaluate_nearfield (i.e., the address of the on-grid history used to correct
the nearfield is NOT the address of the on-grid history that gets FFT'd.)
Thus the FFTs can again be performed in-place. An out-of-place FFT
typically goes faster, but I'm assuming these arrays are going to get HUGE,
so I'm optimizing for memory footprint instead.