The WTF commit.
This commit is _huge_ -- sorry. This brings the nearfield calculation structure
much more in line with that of the direct (AIM) calculation. As the NF matrix
elements aren't time-dependent, they can all be pre-evaluated and cached instead
of doing a BILLION fucking mini-matrix multiplications. This should go much,
much faster but a lot of code had to change to get there.

Major changes:
    * Greatly reduced the scope of AimBase. Because of the revised nearfield,
      the farfield is the only thing to use a lot of these tables.
    * Removed the const & member variables in favor of shared_ptr<const...>
    * AIM::normalization has been renamed to AIM::Normalization
    * Nearfield behaves much more like AIM::DirectInteraction (has a list of
      NF point pairs, stores coefficients in a [pair][time] table, etc.)